pyproject.toml
[project]
name = "upload-batch"
version = "0.1.0"
description = "'сервис по загрузке в MinIO батчей'"
authors = [
    {name = "Ury Dol3wsky",email = "dol3wsky@gmail.com"}
]
license = {text = "\"MIT\""}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[sqlalchemy] (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "minio (>=7.2.15,<8.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pika (>=1.3.2,<2.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.0"

